Refactored code.  Fixed bug with Cell class.

+++ src/main/java/games/sudoku/Reporting.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class Reporting {
 	private int generationAttempts;
-	private long iterations;
+	private int iterations;
 	
 	public void reportGenerationAttempt() {
 		this.generationAttempts++;
